CLI-837 Split telemetry/claude-hooks/binary-refresh migrations out of post-update.ts
post-update.ts bundled four unrelated one-time post-upgrade concerns into a
single 374-line file. Split telemetry sink migration, Claude Code hook
migration, and binary version refresh into their own single-purpose sibling
modules (declarative integration replay was already split out in the prior
commit). post-update.ts now only holds the orchestrator: the version gate,
the four migration calls, and the reload/save/cleanup sequencing.
