+++ tests/unit/core/host/post-update.test.ts
@@ -18,40 +18,23 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-import { randomUUID } from 'node:crypto';
 import * as fs from 'node:fs';
-import { tmpdir } from 'node:os';
-import { join } from 'node:path';
 
 import { afterEach, beforeEach, describe, expect, it, Mock, spyOn } from 'bun:test';
 
 import { supportedIntegrations } from '@/commands/integrate';
 import { CLAUDE_INTEGRATION_ID } from '@/commands/integrate/claude/declaration.ts';
 import * as hooks from '@/commands/integrate/claude/hooks.ts';
-import * as configConstants from '@/core/config-constants.ts';
-import { DISTRIBUTION } from '@/core/host/distribution.ts';
-import * as scaScannerInstall from '@/core/host/install/sca-scanner.ts';
 import * as secretsInstall from '@/core/host/install/secrets.ts';
-import { SCA_SCANNER_BINARY_NAME } from '@/core/host/install-types.ts';
 import * as migration from '@/core/host/migration.ts';
-import {
-  migrateClaudeCodeHooks,
-  migrateLegacyTelemetryEvents,
-  type PostUpdateDependencies,
-  runPostUpdateActions,
-  updateScaScannerBinaryIfNeeded,
-  updateSecretsBinaryIfNeeded,
-} from '@/core/host/post-update.ts';
-import type { CliState, HookExtension, StoredCommandExecutedEvent } from '@/core/state/state.ts';
+import type { PostUpdateDependencies } from '@/core/host/post-update.ts';
+import { runPostUpdateActions } from '@/core/host/post-update.ts';
+import type { CliState } from '@/core/state/state.ts';
 import { getDefaultState } from '@/core/state/state.ts';
 import * as stateRepository from '@/core/state/state-repository.ts';
-import * as telemetryEvents from '@/core/telemetry/telemetry-events.ts';
 import * as versionLib from '@/core/version.ts';
 
 import { version as CURRENT_VERSION } from '../../../../package.json';
-
-const FAKE_HOME = '/fake/home';
-const homedirFn = () => FAKE_HOME;
 
 function makeDeps(): PostUpdateDependencies {
   return {
@@ -63,27 +46,6 @@ function makeDeps(): PostUpdateDependencies {
 
 function makeState(): CliState {
   return getDefaultState('1.0.0');
-}
-
-function makeStateWithExtensions(extensions: HookExtension[], configured = true): CliState {
-  const state = getDefaultState('1.0.0');
-  state.agents['claude-code'].configured = configured;
-  state.agentExtensions = extensions;
-  return state;
-}
-
-function makeExtension(projectRoot: string, global: boolean): HookExtension {
-  return {
-    id: 'test-id',
-    agentId: 'claude-code',
-    kind: 'hook',
-    name: 'sonar-secrets',
-    hookType: 'PreToolUse',
-    projectRoot,
-    global,
-    updatedByCliVersion: '1.0.0',
-    updatedAt: '2026-01-01T00:00:00.000Z',
-  };
 }
 
 describe('runPostUpdateActions', () => {
@@ -222,473 +184,5 @@ describe('runPostUpdateActions', () => {
     expect(saved.agents['claude-code'].hooks.installed.some((h) => h.name === 'sonar-a3s')).toBe(
       false,
     );
-  });
-});
-
-function makeLegacyCommandEvent(command: string): StoredCommandExecutedEvent {
-  return {
-    metadata: {
-      event_id: randomUUID(),
-      source: { domain: 'CLI' },
-      event_type: 'Analytics.Cli.CliCommandExecuted',
-      event_timestamp: String(Date.now()),
-    },
-    event_payload: {
-      cli_installation_id: 'install-id',
-      machine_id: 'machine-id',
-      cli_version: '1.0.0',
-      invocation_id: 'inv-id',
-      os: 'linux',
-      connection_type: null,
-      user_uuid: null,
-      organization_uuid_v4: null,
-      sqs_installation_id: null,
-      caller_agent: null,
-      command,
-      subcommand: null,
-      result: 'success',
-      distribution: DISTRIBUTION,
-    },
-  };
-}
-
-describe('migrateLegacyTelemetryEvents', () => {
-  let loadStateSpy: Mock<typeof stateRepository.loadState>;
-  let saveStateSpy: Mock<typeof stateRepository.saveState>;
-  let appendTelemetryEventSpy: Mock<typeof telemetryEvents.appendTelemetryEvent>;
-  let getTelemetryDirSpy: Mock<typeof configConstants.getTelemetryDir>;
-  let telemetryDir: string;
-
-  beforeEach(() => {
-    telemetryDir = fs.mkdtempSync(join(tmpdir(), 'cli-post-update-telemetry-'));
-    loadStateSpy = spyOn(stateRepository, 'loadState').mockReturnValue(makeState());
-    saveStateSpy = spyOn(stateRepository, 'saveState').mockImplementation(() => {});
-    appendTelemetryEventSpy = spyOn(telemetryEvents, 'appendTelemetryEvent').mockImplementation(
-      () => {},
-    );
-    getTelemetryDirSpy = spyOn(configConstants, 'getTelemetryDir').mockReturnValue(telemetryDir);
-  });
-
-  afterEach(() => {
-    loadStateSpy.mockRestore();
-    saveStateSpy.mockRestore();
-    appendTelemetryEventSpy.mockRestore();
-    getTelemetryDirSpy.mockRestore();
-    fs.rmSync(telemetryDir, { recursive: true, force: true });
-  });
-
-  it('does nothing when there are no legacy telemetry events', () => {
-    // Default state has no telemetry.events queue.
-    migrateLegacyTelemetryEvents();
-
-    expect(appendTelemetryEventSpy).not.toHaveBeenCalled();
-    expect(saveStateSpy).not.toHaveBeenCalled();
-  });
-
-  it('migrates each legacy event to telemetry-events.ndjson and clears the queue', () => {
-    const state = makeState();
-    const events = [makeLegacyCommandEvent('auth'), makeLegacyCommandEvent('analyze')];
-    state.telemetry.events = events;
-    loadStateSpy.mockReturnValue(state);
-
-    migrateLegacyTelemetryEvents();
-
-    expect(appendTelemetryEventSpy).toHaveBeenCalledTimes(2);
-    expect(appendTelemetryEventSpy).toHaveBeenNthCalledWith(1, events[0]);
-    expect(appendTelemetryEventSpy).toHaveBeenNthCalledWith(2, events[1]);
-
-    expect(saveStateSpy).toHaveBeenCalledTimes(1);
-    expect(saveStateSpy.mock.calls[0][0].telemetry.events).toBeUndefined();
-  });
-
-  it('renames the on-disk findings.ndjson sink to telemetry-events.ndjson', () => {
-    const oldPath = join(telemetryDir, 'findings.ndjson');
-    const newPath = join(telemetryDir, 'telemetry-events.ndjson');
-    fs.writeFileSync(oldPath, '{"event":"buffered"}\n');
-
-    migrateLegacyTelemetryEvents();
-
-    expect(fs.existsSync(oldPath)).toBe(false);
-    expect(fs.readFileSync(newPath, 'utf-8')).toBe('{"event":"buffered"}\n');
-  });
-
-  it('appends the old sink contents when telemetry-events.ndjson already exists', () => {
-    const oldPath = join(telemetryDir, 'findings.ndjson');
-    const newPath = join(telemetryDir, 'telemetry-events.ndjson');
-    fs.writeFileSync(newPath, '{"event":"existing"}\n');
-    fs.writeFileSync(oldPath, '{"event":"buffered"}\n');
-
-    migrateLegacyTelemetryEvents();
-
-    expect(fs.existsSync(oldPath)).toBe(false);
-    expect(fs.readFileSync(newPath, 'utf-8')).toBe('{"event":"existing"}\n{"event":"buffered"}\n');
-  });
-});
-
-describe('migrateClaudeCodeHooks', () => {
-  let existsSyncSpy: Mock<typeof fs.existsSync>;
-  let loadStateSpy: Mock<typeof stateRepository.loadState>;
-  let migrateHookScriptsSpy: Mock<typeof migration.migrateHookScripts>;
-  let removeObsoleteHookArtifactsSpy: Mock<typeof migration.removeObsoleteHookArtifacts>;
-  let installHooksSpy: Mock<typeof hooks.installHooks>;
-
-  beforeEach(() => {
-    existsSyncSpy = spyOn(fs, 'existsSync').mockReturnValue(false);
-    loadStateSpy = spyOn(stateRepository, 'loadState').mockReturnValue(makeState());
-    migrateHookScriptsSpy = spyOn(migration, 'migrateHookScripts').mockImplementation(() => {});
-    removeObsoleteHookArtifactsSpy = spyOn(
-      migration,
-      'removeObsoleteHookArtifacts',
-    ).mockResolvedValue(undefined);
-    installHooksSpy = spyOn(hooks, 'installHooks').mockResolvedValue(undefined);
-  });
-
-  afterEach(() => {
-    existsSyncSpy.mockRestore();
-    loadStateSpy.mockRestore();
-    migrateHookScriptsSpy.mockRestore();
-    removeObsoleteHookArtifactsSpy.mockRestore();
-    installHooksSpy.mockRestore();
-  });
-
-  it('does not install hooks when agent is not configured and registry is empty', async () => {
-    loadStateSpy.mockReturnValue(makeState()); // configured = false, no extensions
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).not.toHaveBeenCalled();
-  });
-
-  it('does not install hooks when agent is configured but registry is empty and no global hooks dir exists', async () => {
-    const state = makeStateWithExtensions([]); // configured, no extensions
-    loadStateSpy.mockReturnValue(state);
-    existsSyncSpy.mockReturnValue(false); // globalHooksDir does not exist
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).not.toHaveBeenCalled();
-  });
-
-  it('does not install hooks when registry contains only skill extensions', async () => {
-    const state = makeStateWithExtensions([]);
-    state.agentExtensions = [
-      {
-        id: 'skill-id',
-        agentId: 'claude-code',
-        kind: 'skill',
-        name: 'sonar-context-augmentation',
-        projectRoot: '/some/project',
-        global: false,
-        updatedByCliVersion: '1.0.0',
-        updatedAt: '2026-01-01T00:00:00.000Z',
-      },
-    ];
-    loadStateSpy.mockReturnValue(state);
-    existsSyncSpy.mockReturnValue(false); // global hooks dir does not exist
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).not.toHaveBeenCalled();
-    expect(migrateHookScriptsSpy).not.toHaveBeenCalled();
-  });
-
-  it('skips legacy migration when Claude is already tracked declaratively', async () => {
-    const state = makeStateWithExtensions([makeExtension('/proj/root', false)]);
-    state.integrations.installed.push({
-      id: 'claude-integration-id',
-      integrationId: 'claude-code',
-      installedByCliVersion: '1.0.0',
-      installedAt: '2026-01-01T00:00:00.000Z',
-      updatedByCliVersion: '1.0.0',
-      updatedAt: '2026-01-01T00:00:00.000Z',
-      features: [
-        {
-          featureId: 'mcp-server',
-          scope: 'project',
-          targetRoot: '/proj/root',
-          installedByCliVersion: '1.0.0',
-          installedAt: '2026-01-01T00:00:00.000Z',
-          updatedByCliVersion: '1.0.0',
-          updatedAt: '2026-01-01T00:00:00.000Z',
-          dependencies: [],
-          resources: [],
-          operations: [],
-        },
-      ],
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).not.toHaveBeenCalled();
-    expect(migrateHookScriptsSpy).not.toHaveBeenCalled();
-  });
-
-  it('does not skip legacy migration for an empty declarative Claude container', async () => {
-    const state = makeStateWithExtensions([makeExtension('/proj/root', false)]);
-    state.integrations.installed.push({
-      id: 'claude-integration-id',
-      integrationId: 'claude-code',
-      installedByCliVersion: '1.0.0',
-      installedAt: '2026-01-01T00:00:00.000Z',
-      updatedByCliVersion: '1.0.0',
-      updatedAt: '2026-01-01T00:00:00.000Z',
-      features: [],
-    });
-    loadStateSpy.mockReturnValue(state);
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).toHaveBeenCalledTimes(1);
-    expect(migrateHookScriptsSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('installs hooks for each extension in the registry', async () => {
-    const state = makeStateWithExtensions([makeExtension('/proj/root', false)]);
-    loadStateSpy.mockReturnValue(state);
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('passes projectRoot and undefined globalDir for non-global extensions', async () => {
-    const state = makeStateWithExtensions([makeExtension('/proj/root', false)]);
-    loadStateSpy.mockReturnValue(state);
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).toHaveBeenCalledWith('/proj/root', undefined, false);
-  });
-
-  it('passes projectRoot and homedirFn() as globalDir for global extensions', async () => {
-    const state = makeStateWithExtensions([makeExtension('/proj/root', true)]);
-    loadStateSpy.mockReturnValue(state);
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).toHaveBeenCalledWith('/proj/root', FAKE_HOME, false);
-  });
-
-  it('migrates hook scripts for each location before installing hooks', async () => {
-    const state = makeStateWithExtensions([makeExtension('/proj/root', false)]);
-    loadStateSpy.mockReturnValue(state);
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(migrateHookScriptsSpy).toHaveBeenCalledTimes(1);
-    expect(migrateHookScriptsSpy).toHaveBeenCalledWith('/proj/root', undefined);
-  });
-
-  it('deduplicates locations - installs hooks once for repeated (projectRoot, globalDir)', async () => {
-    const state = makeStateWithExtensions([
-      makeExtension('/proj/root', false),
-      makeExtension('/proj/root', false),
-    ]);
-    loadStateSpy.mockReturnValue(state);
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('installs hooks for multiple distinct locations', async () => {
-    const state = makeStateWithExtensions([
-      makeExtension('/proj/alpha', false),
-      makeExtension('/proj/beta', false),
-    ]);
-    loadStateSpy.mockReturnValue(state);
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).toHaveBeenCalledTimes(2);
-  });
-
-  it('falls back to global migration when registry is empty, agent is configured, and global hooks dir exists', async () => {
-    const state = makeStateWithExtensions([]); // configured, no extensions
-    loadStateSpy.mockReturnValue(state);
-    existsSyncSpy.mockReturnValue(true); // globalHooksDir exists
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('uses homedirFn() as both projectRoot and globalDir in the pre-registry fallback', async () => {
-    const state = makeStateWithExtensions([]);
-    loadStateSpy.mockReturnValue(state);
-    existsSyncSpy.mockReturnValue(true);
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).toHaveBeenCalledWith(FAKE_HOME, FAKE_HOME, false);
-  });
-
-  it('does not fall back when agent is not configured', async () => {
-    const state = makeStateWithExtensions([], false); // configured = false
-    loadStateSpy.mockReturnValue(state);
-    existsSyncSpy.mockReturnValue(true); // hooks dir exists, but shouldn't matter
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(installHooksSpy).not.toHaveBeenCalled();
-  });
-
-  it('continues installing remaining locations when one throws', async () => {
-    const state = makeStateWithExtensions([
-      makeExtension('/proj/alpha', false),
-      makeExtension('/proj/beta', false),
-    ]);
-    loadStateSpy.mockReturnValue(state);
-    migrateHookScriptsSpy.mockImplementationOnce(() => {
-      throw new Error('migrate failed');
-    });
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    // First location failed, but second location still ran
-    expect(installHooksSpy).toHaveBeenCalledTimes(1);
-    expect(installHooksSpy).toHaveBeenCalledWith('/proj/beta', undefined, false);
-  });
-
-  it('does not throw when a location migration fails', async () => {
-    const state = makeStateWithExtensions([makeExtension('/proj/root', false)]);
-    loadStateSpy.mockReturnValue(state);
-    installHooksSpy.mockRejectedValue(new Error('hook install failed'));
-
-    const actual = await migrateClaudeCodeHooks(
-      hooks.installHooks,
-      CLAUDE_INTEGRATION_ID,
-      homedirFn,
-    );
-
-    expect(actual).toBeUndefined();
-  });
-
-  it('calls removeObsoleteHookArtifacts once per location with the sonar-a3s marker', async () => {
-    const state = makeStateWithExtensions([
-      makeExtension('/proj/alpha', false),
-      makeExtension('/proj/beta', false),
-    ]);
-    loadStateSpy.mockReturnValue(state);
-
-    await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
-
-    expect(removeObsoleteHookArtifactsSpy).toHaveBeenCalledTimes(2);
-    expect(removeObsoleteHookArtifactsSpy).toHaveBeenCalledWith(
-      '/proj/alpha',
-      migration.OBSOLETE_A3S_MARKER,
-    );
-    expect(removeObsoleteHookArtifactsSpy).toHaveBeenCalledWith(
-      '/proj/beta',
-      migration.OBSOLETE_A3S_MARKER,
-    );
-  });
-});
-
-function makeStateWithSecrets(): CliState {
-  const state = makeState();
-  state.tools = {
-    installed: [
-      {
-        name: 'sonar-secrets',
-        version: '0.0.0.1',
-        path: '/fake/bin/sonar-secrets-0.0.0.1-linux-x86-64',
-        installedAt: '2026-01-01T00:00:00.000Z',
-        installedByCliVersion: '1.0.0',
-      },
-    ],
-  };
-  return state;
-}
-
-describe('updateSecretsBinaryIfNeeded', () => {
-  let loadStateSpy: Mock<typeof stateRepository.loadState>;
-  let installSecretsBinarySpy: Mock<typeof secretsInstall.installSecretsBinary>;
-
-  beforeEach(() => {
-    loadStateSpy = spyOn(stateRepository, 'loadState').mockReturnValue(makeStateWithSecrets());
-    installSecretsBinarySpy = spyOn(secretsInstall, 'installSecretsBinary').mockResolvedValue(
-      '/fake/bin/sonar-secrets',
-    );
-  });
-
-  afterEach(() => {
-    loadStateSpy.mockRestore();
-    installSecretsBinarySpy.mockRestore();
-  });
-
-  it('does nothing when no previous binary is recorded in state', async () => {
-    loadStateSpy.mockReturnValue(makeState()); // tools.installed is empty
-
-    await updateSecretsBinaryIfNeeded();
-
-    expect(installSecretsBinarySpy).not.toHaveBeenCalled();
-  });
-
-  it('calls installSecretsBinary when a previous installation is recorded in state', async () => {
-    await updateSecretsBinaryIfNeeded();
-
-    expect(installSecretsBinarySpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('propagates errors from installSecretsBinary to the caller', () => {
-    installSecretsBinarySpy.mockRejectedValue(new Error('download failed'));
-
-    expect(updateSecretsBinaryIfNeeded()).rejects.toThrow('download failed');
-  });
-});
-
-function makeStateWithScaScanner(): CliState {
-  const state = makeState();
-  state.tools = {
-    installed: [
-      {
-        name: SCA_SCANNER_BINARY_NAME,
-        version: '0.0.0.1',
-        path: '/fake/bin/sca-scanner-cli-0.0.0.1-linux-x86-64',
-        installedAt: '2026-01-01T00:00:00.000Z',
-        installedByCliVersion: '1.0.0',
-      },
-    ],
-  };
-  return state;
-}
-
-describe('updateScaScannerBinaryIfNeeded', () => {
-  let loadStateSpy: Mock<typeof stateRepository.loadState>;
-  let installScaScannerBinarySpy: Mock<typeof scaScannerInstall.installScaScannerBinary>;
-
-  beforeEach(() => {
-    loadStateSpy = spyOn(stateRepository, 'loadState').mockReturnValue(makeStateWithScaScanner());
-    installScaScannerBinarySpy = spyOn(
-      scaScannerInstall,
-      'installScaScannerBinary',
-    ).mockResolvedValue('/fake/bin/sca-scanner-cli');
-  });
-
-  afterEach(() => {
-    loadStateSpy.mockRestore();
-    installScaScannerBinarySpy.mockRestore();
-  });
-
-  it('does nothing when no previous binary is recorded in state', async () => {
-    loadStateSpy.mockReturnValue(makeState()); // tools.installed is empty
-
-    await updateScaScannerBinaryIfNeeded();
-
-    expect(installScaScannerBinarySpy).not.toHaveBeenCalled();
-  });
-
-  it('calls installScaScannerBinary when a previous installation is recorded in state', async () => {
-    await updateScaScannerBinaryIfNeeded();
-
-    expect(installScaScannerBinarySpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('propagates errors from installScaScannerBinary to the caller', () => {
-    installScaScannerBinarySpy.mockRejectedValue(new Error('download failed'));
-
-    expect(updateScaScannerBinaryIfNeeded()).rejects.toThrow('download failed');
   });
 });
